Silence UBSAN on empty vector_view when parsing source code
The fix is just a band-aid, though. The best thing you can do is to
avoid constructing empty views, as trying to make them make sense is a
quick way to Insanityville.

What will front() return? Undefined behaviour.

What will back() return? Undefined behaviour.

What about data()? Same.

Is size() safe? Nope, because without special casing the empty view we
will use std::distance() on a pair of nullptrs and that will probably
not end well.

So yeah, silence the sanitiser, but avoid doing anything with the empty
views except for iterating over them and checking for emptiness.

Curiously, when I added a check:

    if (empty()) {
        throw std::logic_error{"empty view"};
    }

to front() and back() GCC started complaining that

src/tools/exec/asm.cpp:953:41:   required from here
  952 |     while (not lexemes.empty()) {
  953 |         auto const& each = lexemes.front();
      |                            ~~~~~~~~~~~~~^~
./include/viua/support/vector.h:105:13: error: ‘throw’ will always call ‘terminate’ [-Werror=terminate]
  105 |             throw std::logic_error{"viua::support::vector_view::front on empty view" };

which is interesting.

I could just say -Wno-error=terminate, but I want to keep compile logs
as clean as possible and I want to avoid conditioning myself to ignore
warnings because "it is probably that thing was OK" and having something
sneak up on me. Remember, only the paranoid survive.

That is why I not only special cased the empty view, but also added the
check for it in assembler's source.