Add a "set" kind of option to command line processing code
This kind is similar to the "list" one, but does not care about ordering
of items in the sequence. It is useful for flags eg,

    ]$ viua asm --dump lex-raw --dump ast foo.asm

It does not matter in what order the --dump options were given; what is
important is whather a value is present in the set or not.