Make toolchain frontend more flexible vis a vis tool selection
If something is not detected as an option or a built-in tool, then the
command is simply forwarded to the tool-invoking function. That
function, in turn, is responsible for reporting a possible error when
the selected tool does not exist.

This adds more flexibility to the frontend, and allows "side loading"
tools ad-hoc.