Binding and visibility (version 2)
I know what I want to say. I am just not sure how to best express and
present the ideas. It is pretty simple: Viua uses ELF's binding and
visibility to create four kinds of symbols in the .text section:

 - jump targets
 - compilation unit-private functions
 - module-private functions
 - public functions

and two kinds of symbols in the .rodata section:

 - compilation unit-private data
 - module-private data

There are no public ie, accessible outside their compilation unit or
module, variables in Viua.

Now, what is a "compilation unit" and what is a "module"?

By "compilation unit" I mean a file. The assembler takes a single file
as input, compiles it, and produces a single file as output.

By "module" I mean whatever the linker is used to produce. A module can
be composed of several compilation units eg,

    ]$ viua asm -o part1.o part1.s
    ]$ viua asm -o part2.o part2.s
    ]$ viua ld --type shared -o libpart.so part1.o part2.o

In the example below part*.o are compilation units, and libpart.so is a
module. It only works like this during the FIRST time the linker is run
on an object file. Why?

Because during the first time the linker processes an object file it
patches any items in the .symtab with global binding (STB_GLOBAL) and
hidden visibility (STV_HIDDEN) to have local binding (STB_LOCAL) and
default visibility (STV_DEFAULT), making them not available outside of
their local scope ie, the ELF file they are contained in.