Prepare to document instructions
I think I will move the detailed documentation into viua-asm(7), because
the page in section 5 should deal with the filr format—not instruction
semantics.

Then why did I document directives in viua-asm(5)? Because it would be
useless to say something like

    A directive is a period ie, the "." character, followed by a
    sequence of [a-z] characters; it may take an optional operand. An
    operand is either a string (enclosed in double quotes eg, "foo"), a
    sequence of [a-z] characters, or an expression specific to a
    particular directive (consult detailed documentation of that
    specific directive on the viua-asm(7) man page).

Sure, it tells me what a directive is and what it looks like, but it is,
at the same time, utterly useless. Maybe the discussion of ELF semantics
could be moved to section 7, but I am not sure whether it makes sense to
split it this way.

Detailed instruction documentation? It absolutely makes sense to move it
to the viua-asm(7) page. There are many instructions, and it would be
useful to be able to give each category its own section and have eg, IO,
ARITHMETIC, and CONTROL FLOW sections, with subsections dedicated to
instructions (or groups of instructions).

The viua-asm(5) will concern itself with documenting _formats_ of
instructions ie, the N, T, D, S, F, E, R, and M type of an operation and
show their general form. Then, viua-asm(7) will be free to discuss the
semantics without delving into syntax.

What about types? Some operations are type-agnostic, but some rely on
the virtual machine being aware of what is in the register it should
use. I think they should be discusses in the beginning of viua-asm(7),
get their own viua-types(7), or be discussed in viua-isa(7) which would
be concerned with the ISA as a whole: registers, types, flags, etc.