Implement an I/O backend using the "classic" API
The I/O backend is selectable at compile-time with:

    ]$ make io_backend=<backend>

where <backend> is either "classic" or "io_uring".

The "classic" API is based on simple read(2) and write(2) calls, is
inefficient and just plain bad, but should work everywhere. Also, this
is a hobby project so I am going to go with whatever *works* and is not
too annoying, and not strive for absolute technical perfection if it is
not a current goal (and it is not).

The "io_uring" API is based on, well, io_uring.

Now, why would I need the classic API? Because I want the VM to run on
BSDs and there is no io_uring in BSD-land.