Fix requested program interpreter
There is no "viua-vm" program installed anymore, so requesting it as the
interpreter for Viua ELF files does not make any sense.

Since the intepreter is embedded in the ELF information, the proper path
must be known at compile time. Make sure you are compiling and
installing the VM using the same

    ]$ make prefix=foo all
    ]$ make prefix=foo install

prefix= value for both steps.