Allow multiple long labels for a single option
I want to be compatible with the interface of readelf(1) from GNU
Binutils, and to be able to do that I must be able to support multiple
long labels. See readelf(1) to see what I mean, but to give an example:

    readelf [-a|--all]
            [-l|--program-headers|--segments]

The -l option has one short and two long labels.