Add FreeBSD as a supported platform
The VM compiles, runs, and passes tests on FreeBSD.

Porting VM to FreeBSD was not too difficult. It was mostly a matter of:

 - having a recent-enough compiler to support the parts of C++23 that I
   use
 - cutting out SHA2-384 as it was a gimmicky feature since there are
   SHA2-512, BLAKE2b, and BLAKE3 available as build-id hashes
 - hiding BLAKE3 behind a feature test
 - adjusting to BSD userland or explicitly invoking GNU tools

That is not to say that this was a piece of cake and an easy stroll
through a park; it was a piece of stale bread and a walk through a bad
part of town, because FreeBSD is different enough to be annoying. It is
like the interfaces I use are *almost* the same on Linux and BSD, but
not quite... and this adds unnecessary friction.

Forcing myself to support Alpine with musl as well as one of the BSDs is
a good exercise, though. Makes the code better organised and more
flexible. For example, take the Makefile or the I/O backends
reorganisation; both were ultimately good things.