There is no std::observer_ptr anymore
Here is Stroustrup's paper on the issue:
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2019/p1408r0.pdf

It is safe to assume that std::observer_ptr is gone, but I kind of liked
the idea--even if not the name--so this patch introduces a type that is
a drop-in replacement: viua::view_ptr.

As a bonus, viua::view_ptr will throw if you try to access a null
pointer through it, which I consider a good thing. Instead of underfined
or random behaviour you get a BIG LOUD CRASH.