Ignore errors when emitting header dependencies
Not all dependencies can be satisfied on all platforms eg, on FreeBSD
there is no <liburing.h> header so the rule would always fail. Just
ignoring the error in this one particular case is a much simpler
solution than trying to filter out the files that are not supported on
the current platform.