Limit parallel jobs to 2 in CI
This is the number of cores Sourcehut provides to the build VMs, and
having more parallelism than cores does not really give us anything.
Although in this case this is besides the point, really.

The crux of the issue was running make(1) with unbounded -j which led to
the OOM getting triggered by the builds using the GCC compiler. Limiting
number of concurrent jobs fixes the problem.