Rearrange fields inside an encoded instruction
By putting the opcode in the lowest bits we avoid the need for shifting
to get the opcode value out of the instruction.

Putting the immediate in the high word in every format means that the
same shift is required in all cases: 32 bits to the right. Same other
fields - always the same shift is necessary, regardless of the format.