Explicit test cases as arguments not comma-separated
The test suite now takes explicit test cases as a arguments eg,

    ]$ python3 ./test/suite.py foo bar baz

instead of as a comma-separated list in a single argument eg,

    ]$ python3 ./test/suite.py foo,bar,baz

This does not change the functionality of the test suite on iota, but
makes it easier to start it with a custom set of test cases to run. For
example, to glob for cases one can use:

    ]$ python3 ./test/suite.py $(ls -1 ./tests/asm/*.asm |
        xargs -n 1 basename |
        sed 's/\.asm$//')

This was much more cumbersome before, as you would have to somehow
massage the output of ls(1) into a comma-separated list, which was just
a useless complication to overcome.