Rename tests to new scheme
The new scheme has the old prefix (star, for STyled ARithmetic), but
uses a different middle part. Instead of first letter of the style and
the signedness eg, ss for saturating signed, the new scheme uses tags
for styles and explicit bit widths for integer type. For example:

 style      |      integer       |   operation    | test
            | width | signedness |                | name
 -----------+-------+------------+----------------+-------------
 SATurating |   8   | signed     | addition       | sat_i8_add
 -----------+-------+------------+----------------+-------------
 WRaPping   |  16   | unsigned   | subtraction    | wrp_u16_sub
 -----------+-------+------------+----------------+-------------
 TRaPping   |  24   | signed     | multiplication | trp_i24_mul
 -----------+-------+------------+----------------+-------------

I think this scheme is much more obvious than the old one.