Factory functions for arithmetic types
The arithmetic_type only got two factories:

 - of_size() for producing arithmetic values of the requested size
 - zero() for producing zeroes of the requested size

Since the arithmetic type does not have a concept of signedness, it is
neither signed nor unsigned, it does not really make sense to provide it
with factories for its maximum and minimum; the type simply has neither.

The signed_type got three factories:

 - max() for producing values representing the upper limit of the
   requested size
 - min() for producing values representing the lower limit of the
   requested size
 - zero() for producing zeroes of the requested size

Since signed_type is a bit (pun intended) more restricted than the base
arithmetic_type it received the max() and min() factories instead of the
incredibly general of_size() factory. We know how to interpret the bit
patterns of the signed_type so having max() and min() makes sense--they
are immediately useful.

The arithmetic_type::of_size() acts as the low-level building block used
to implement the max() and min() factories.