Saner pseudoinstruction expansion
The current code was an ad-hoc mess. I should have done better the first
time, but it is what it is. The code grows, and sometimes what started
as a nice quiet little park, becomes a forest; the important thing is to
clean it up before it becomes an unmanageable jungle.