Make test suite output fit in 80 columns
Except for the initial "using toolchain in:" report, because I cannot
control the length of the path to the directory where the repository is
cloned.

If there are more than 80 columns, use them for graphs, but But not for
regular output. Graphs can easily scale, but the output would require
selecting different symbols, format strings, etc to handle variable
terminal sizes properly. I do not think it is worth the effort.