Gather data about instruction timings
It is useful to know that an XYZ instruction was hit N times, but it is
not enough to fully assess its impact. The missing piece of information
is how long does it usually take to execute that instruction -- this
patch provides the test suite with this information.

The test suite now reports two main measurements:

 - instruction hit rate ie, how many times an instruction has been
   executed by the tests
 - instruction time averages ie, how long does it take, on average, to
   execute a particular instruction

and a single measure derived from the two above:

 - instruction impact ie, what is the percentage of run time dedicated
   to executing this instruction

The impact is simply hit rate multiplied by the average time, and is
what is displayed by default.

The main measurements are hidden by default, since showing all three
tables would take too much vertical space on the screen; and the impact
table includes hit rate and timings. Detailed reporting can be enabled
by setting the SHOW_PERFORMANCE_DETAILS environment variable to true
before running the test suite:

    ]$ export SHOW_PERFORMANCE_DETAILS=true
    ]$ make test

Keep in mind that if you do so, your screen will be devoured by the
output. Best use it in moderation. (The detailed output is enabled in
CI, because why not.)