Fix viua::support::ston
Now it works for correctly, although the error reports from the
assembler could be better. Consider this instruction:

    li $1.l, 0xffffffffffffffff

The number to be loaded here is "clearly" a 64-bit number, but the
assembler will complain that it must fit into 64 bits. What is going on?

Well, the number represented is signed, and 0xffffffffffffffff cannot
fit into a SIGNED 64-bit integer. You have to write the instruction as

    li $1.l, -1

to get your desired value. The assembler could try to be more
user-friendly, but I think the current message will do for now.