Better naming scheme for styled arithmetic tests
The new scheme is not necessarily more consistent or anything, but it
will make the test suite output more readable. Imagine, if you will, two
test cases:

 - for wrapping addition of signed 8-bit integers
 - for saturating multiplication of unsigned 13-bit integers

Using the old scheme the names would be

 - arst_wrp_i8_add
 - arst_sat_u13_mul

It is clear that there is a little bit of irregularity here: the name of
the operation is not aligned, even though it could be - after all, both
"add" and "mul" have the same width! The new scheme makes the names
align

 - arst_wrp_add_i8
 - arst_sat_mul_u13

and it even follows the order of the plain-English version of the name:

 - arst: test of the styled arithmetic instruction...
 - wrp: for wrapping...
 - add: addition...
 - i8: of 8-bit signed integers

Same information conveyed, but in a better way!