Fix test suite crashing when a requested test case is missing
This is embarassing, but such is life. How did I find the crash? By
having an arst_sat_sub_u8.ebreak file without matching .asm file, and
running the following command:

    ]$ bash ./scripts/watch-some-tests tests/asm/arst_*.{asm,ebreak}

The shell picked up the .ebreak and requested the test suite to run the
matching case... which made the suite explode.