Unsigned division
This commit completes the implementation of the two most important
arithmetic styles, fixed and saturating. The implementation is based on
naive (ie, correct but slow) algorithms and is meant to be easy to
maintain, not to be the most performant one.

The goal was to provide a base against which tests could have been
written. Now that this has been accomplished, the implementation can be
replaced with a more performant one. The test suite will ensure that no
bugs sneak in.