Add floating-point arithmetic tests
I do not have interesting numbers for them, but it is not important,
since floating-point relies completely on the underlying hardware
platform - testing edge cases has been handled lower on the stack.

What is interesting on the VM level is proper encoding and decoding of
instructions, loading of operands (since for doubles there is some
memory interaction), and the general "handling" of floating points.