Introduce the MOVEIF instruction to replace CMP
CMP can be implemented in terms of EQ, LT, and GT, and is therefore
redundant. However, without MOVEIF the instruction sequence required to
produce a "synthetic CMP" is not as nice as with MOVEIF available.

Of course, even with MOVEIF the code is not exactly pretty, but it is
servicable and easy to understand. To wit:

    cmp $result, $lhs, $rhs

can be written as:

    ; Check if $lhs < $rhs. The multiplication is necessary because lt
    ; produces either a 0 or 1, but we want a -1.
    lt $lt, $lhs, $rhs
    muli $lt, $lt, -1

    ; Check if $lhs > $rhs. No post-processing is necessary, as in this
    ; case we want a 0 or 1, and this is what gt gives us.
    gt $gt, $lhs, $rhs

    ; Move $lt into $result if $lt is non-zero (ie, if the lhs was
    ; less-than rhs); otherwise move $gt into $result. This gives us the
    ; desired result of -1 on less-than, 0 on equals, and 1 on
    ; greater-than relationship between lhs and rhs.
    moveif $result, $lt, $gt

    ; Convert the unsigned $result to a signed value.
    add $result, zero, $result

Simple.