Implement sign() using MOVEIF instructions
This is a bit more straightforward than the old if-based version. Or is
it? I guess it is as long as the comments are there, but... I also
commented the version with ifs, so it is a draw.

In either case, there are comparisons somewhere. The difference between
the if- and moveif-based implementations is whether the flow branches or
not.